Version 0.0.0
A shell of the app. Has support for account creation, login, and navigation to the main user screen. The .csv file is needed for local login storage. In the production version, this will live on our hosting/database machine, and the two functions that communicate with it will run only on the host machine, so the only thing the user system gets back is a string that tells which screen to switch to (the user then should have no way to re-engineer the app the steal passwords, etc). Naming conventions have been set in the app - please follow them. The code here should contain Kivy code for all the main functionality of the app: switching screens with buttons, taking in text, and others.